fix: [ROUND-11b] Broker units float + 50% risk reduction step + discretionary
BROKER (3 bugs de audit agent):
- OrderResult.units, TradeInfo.units: int→float (crypto fractional sizes truncaban a 0)
- place_limit_order, place_stop_order: units int→float (Capital.com soporta fracciones)
- _confirm_deal: units int→float (consistencia con place_market_order)

POSITION MANAGEMENT (mentoría transcription — nueva fase intermedia):
- Añadido paso de 50% reducción de riesgo ANTES del Break Even
  Mentoría: "antes del break-even, al 50% de beneficio, muevo el SL para eliminar
  el 50% del riesgo" — reducción progresiva antes del BE completo
  Secuencia: SL original → 50% risk reduction → Break Even → EMA trailing

Round 11 hallazgos de 2 agentes:
- Broker audit: 5 bugs (units type, circuit breaker bypass, stop order retry)
- Transcription audit: 12 gaps (strategy details, 50% risk step, run vs pullback)
- 3 corregidos este commit, resto tracked para futuras rondas

45 tests pass, 0 TypeScript errors.

+++ backend/broker/base.py
@@ -46,7 +46,7 @@ class OrderResult:
     success: bool
     trade_id: Optional[str] = None
     fill_price: Optional[float] = None
-    units: int = 0
+    units: float = 0
     error: Optional[str] = None
     raw_response: Optional[Dict] = None
 
@@ -69,7 +69,7 @@ class TradeInfo:
     trade_id: str
     instrument: str
     direction: str  # "BUY" or "SELL"
-    units: int
+    units: float
     entry_price: float
     current_price: float
     unrealized_pnl: float
@@ -126,7 +126,7 @@ class BaseBroker(ABC):
     async def place_market_order(
         self,
         instrument: str,
-        units: int,
+        units: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
     ) -> OrderResult:
@@ -137,7 +137,7 @@ class BaseBroker(ABC):
     async def place_limit_order(
         self,
         instrument: str,
-        units: int,
+        units: float,
         price: float,
         stop_loss: Optional[float] = None,
         take_profit: Optional[float] = None,
